Extend OLCand REST API to allow not to create products on the fly
* fix unit test
* move a bunch of static *String* related methods from Util to StringUtils

https://github.com/metasfresh/metasfresh/issues/4828
