+++ de.metas.adempiere.adempiere/base/src/main/java-legacy/org/compiere/print/MPrintFormat.java
@@ -39,7 +39,6 @@ import org.compiere.model.X_AD_PrintFormat;
 import org.compiere.util.CPreparedStatement;
 import org.compiere.util.DB;
 import org.compiere.util.Env;
-import org.compiere.util.Util;
 import org.slf4j.Logger;
 
 import de.metas.cache.CCache;
@@ -47,6 +46,7 @@ import de.metas.cache.interceptor.CacheCtxParamDescriptor;
 import de.metas.i18n.Language;
 import de.metas.i18n.Msg;
 import de.metas.logging.LogManager;
+import de.metas.util.StringUtils;
 
 /**
  *	AD_PrintFormat - Print Format Model.
@@ -58,7 +58,7 @@ import de.metas.logging.LogManager;
 public class MPrintFormat extends X_AD_PrintFormat
 {
 	/**
-	 * 
+	 *
 	 */
 	private static final long serialVersionUID = 3626220385155526700L;
 
@@ -82,7 +82,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 			setIsDefault(false);
 		}
 	}	//	MPrintFormat
-	
+
 	/**
 	 * 	Load Constructor
 	 *	@param ctx context
@@ -112,10 +112,10 @@ public class MPrintFormat extends X_AD_PrintFormat
 	protected void onLoadComplete (final boolean success)
 	{
 		// Reset cached objects
-		m_items = null; 
+		m_items = null;
 		m_tFormat = null;
 	}   //  loadComplete
-	
+
 	/**
 	 * 	Get Language
 	 *  @return language
@@ -209,7 +209,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		}
 		return m_items;
 	}
-	
+
 	private MPrintFormatItem[] retrieveItems()
 	{
 		final String whereClause = "AD_PrintFormat_ID=?"
@@ -217,12 +217,12 @@ public class MPrintFormat extends X_AD_PrintFormat
 			+ " AND NOT EXISTS (SELECT * FROM AD_Field f "
 				+ "WHERE AD_PrintFormatItem.AD_Column_ID=f.AD_Column_ID"
 				+ " AND (f.IsEncrypted='Y' OR f.ObscureType IS NOT NULL))";
-		
+
 		final List<MPrintFormatItem> list = new Query(getCtx(), I_AD_PrintFormatItem.Table_Name, whereClause, get_TrxName())
 				.setParameters(getAD_PrintFormat_ID())
 				.setOrderBy(I_AD_PrintFormatItem.COLUMNNAME_SeqNo)
 				.list(MPrintFormatItem.class);
-		
+
 		final IUserRolePermissions role = Env.getUserRolePermissions(getCtx());
 		for(Iterator<MPrintFormatItem> it = list.iterator(); it.hasNext();)
 		{
@@ -232,7 +232,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 				it.remove();
 			}
 		}
-		
+
 		//
 		MPrintFormatItem[] retValue = new MPrintFormatItem[list.size()];
 		list.toArray(retValue);
@@ -276,7 +276,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		log.debug("setTranslation #" + no);
 	}	//	setTranslation
 
-	
+
 	/**************************************************************************
 	 * 	Set Standard Header
 	 *	@param standardHeaderFooter true if std header
@@ -304,7 +304,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 			super.setIsForm(false);
 	}	//	setIsTableBased
 
-	
+
 	/**************************************************************************
 	 * 	Set Translation View Language.
 	 * 	@param language language (checked for base language)
@@ -348,7 +348,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		}
 	}	//	setTranslationViewQuery
 
-	
+
 	/**************************************************************************
 	 * 	Get Optional TableFormat
 	 * 	@param AD_PrintTableFormat_ID table format
@@ -388,7 +388,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		return sb.toString();
 	}	//	toString
 
-	
+
 	/**************************************************************************
 	 *  Load Special data (images, ..).
 	 *  To be extended by sub-classes
@@ -444,7 +444,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 	 *  @param AD_PrintFormat_ID 0 or existing PrintFormat
 	 * 	@return print format
 	 */
-	static public MPrintFormat createFromTable (Properties ctx, 
+	static public MPrintFormat createFromTable (Properties ctx,
 		int AD_Table_ID, int AD_PrintFormat_ID)
 	{
 		int AD_Client_ID = Env.getAD_Client_ID(ctx);
@@ -681,13 +681,13 @@ public class MPrintFormat extends X_AD_PrintFormat
 			catch (SQLException e)
 			{
 				s_log.error("(table) - " + sql, e);
-			}			
+			}
 			finally {
 				DB.close(rs, pstmt);
 				rs = null; pstmt = null;
 			}
 		}
-		
+
 		//
 		MPrintFormatItem[] retValue = new MPrintFormatItem[list.size()];
 		list.toArray(retValue);
@@ -725,18 +725,18 @@ public class MPrintFormat extends X_AD_PrintFormat
      *	@param fromItems from items
      *	@param toItems to items
      */
-    static private void copyTranslationItems (MPrintFormatItem[] fromItems, 
+    static private void copyTranslationItems (MPrintFormatItem[] fromItems,
     	MPrintFormatItem[] toItems)
     {
     	if (fromItems == null || toItems == null)
             return;		//	should not happen
-        
+
     	int counter = 0;
         for (int i = 0; i < fromItems.length; i++)
-        {   
+        {
             int fromID = fromItems[i].getAD_PrintFormatItem_ID();
             int toID = toItems[i].getAD_PrintFormatItem_ID();
-            
+
             StringBuffer sql = new StringBuffer("UPDATE AD_PrintFormatItem_Trl new ")
             	//	Set
             	.append("SET (PrintName, PrintNameSuffix, IsTranslated) = ")
@@ -761,7 +761,7 @@ public class MPrintFormat extends X_AD_PrintFormat
         s_log.trace("#" + counter);
     }	//	copyTranslationItems
 
-	
+
 	/**************************************************************************
 	 * 	Copy existing Definition To Client
 	 * 	@param ctx context
@@ -769,7 +769,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 	 * 	@param to_AD_PrintFormat_ID format
 	 * 	@return print format
 	 */
-	public static MPrintFormat copy (Properties ctx, 
+	public static MPrintFormat copy (Properties ctx,
 		int from_AD_PrintFormat_ID, int to_AD_PrintFormat_ID)
 	{
 		return copy (ctx, from_AD_PrintFormat_ID, to_AD_PrintFormat_ID, -1);
@@ -782,7 +782,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 	 * 	@param to_Client_ID to client
 	 * 	@return print format
 	 */
-	public static MPrintFormat copyToClient (Properties ctx, 
+	public static MPrintFormat copyToClient (Properties ctx,
 		int AD_PrintFormat_ID, int to_Client_ID)
 	{
 		return copy (ctx, AD_PrintFormat_ID, 0, to_Client_ID);
@@ -800,7 +800,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		int to_AD_PrintFormat_ID, int to_Client_ID)
 	{
 		s_log.info("From AD_PrintFormat_ID=" + from_AD_PrintFormat_ID
-			+ ", To AD_PrintFormat_ID=" + to_AD_PrintFormat_ID 
+			+ ", To AD_PrintFormat_ID=" + to_AD_PrintFormat_ID
 			+ ", To Client_ID=" + to_Client_ID);
 		if (from_AD_PrintFormat_ID == 0)
 			throw new IllegalArgumentException ("From_AD_PrintFormat_ID is 0");
@@ -816,9 +816,9 @@ public class MPrintFormat extends X_AD_PrintFormat
 			to.setClientOrg (to_Client_ID, 0);
 		}
 		//	Set Name - Remove TEMPLATE - add copy
-		to.setName(Util.replace(to.getName(), "TEMPLATE", String.valueOf(to_Client_ID)));
-		to.setName(to.getName() 
-			+ " " + Msg.getMsg(ctx, "Copy")		 
+		to.setName(StringUtils.replace(to.getName(), "TEMPLATE", String.valueOf(to_Client_ID)));
+		to.setName(to.getName()
+			+ " " + Msg.getMsg(ctx, "Copy")
 			+ " " + to.hashCode());		//	unique name
 		//
 		to.save();
@@ -847,7 +847,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		if (!readFromDisk)
 		{
 			printFormat = s_formats.get(key);
-			
+
 			// Validate the context
 			if (printFormat != null && !CacheCtxParamDescriptor.isSameCtx(ctx, printFormat.getCtx()))
 			{
@@ -862,7 +862,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 			else
 				s_formats.put(key, printFormat);
 		}
-		
+
 		//
 		// Return a copy
 		if (printFormat != null)
@@ -889,7 +889,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 			sql += "AD_ReportView_ID=?";
 		else
 			sql += "AD_Table_ID=?";
-		sql += " ORDER BY IsDefault DESC"; 
+		sql += " ORDER BY IsDefault DESC";
 		try
 		{
 			pstmt = DB.prepareStatement (sql, null);
@@ -918,7 +918,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		Integer key = new Integer(AD_PrintFormat_ID);
 		s_formats.put(key, null);
 	}	//	deleteFromCache
-	
+
     //begin vpj-cd e-evolution
 	/**
 	 * Get ID of Print Format use Name
@@ -934,10 +934,10 @@ public class MPrintFormat extends X_AD_PrintFormat
 		return DB.getSQLValue(null, sql, formatName, AD_Table_ID, AD_Client_ID);
 	}
 	//end vpj-cd e-evolution
-	
+
 	/**
 	 * @param AD_Table_ID
-	 * @param AD_Client_ID use -1 to retrieve from all client 
+	 * @param AD_Client_ID use -1 to retrieve from all client
 	 * @param trxName
 	 */
 	public static RowSet getAccessiblePrintFormats (int AD_Table_ID, int AD_Client_ID, String trxName)
@@ -950,7 +950,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 		{
 			sql = sql + " AND AD_Client_ID = ? ";
 		}
-		sql = sql + "ORDER BY AD_Client_ID DESC, IsDefault DESC, Name"; //	Own First 
+		sql = sql + "ORDER BY AD_Client_ID DESC, IsDefault DESC, Name"; //	Own First
 		//
 		sql = Env.getUserRolePermissions().addAccessSQL (
 			sql, "AD_PrintFormat", IUserRolePermissions.SQL_NOTQUALIFIED, IUserRolePermissions.SQL_RO);
@@ -971,7 +971,7 @@ public class MPrintFormat extends X_AD_PrintFormat
 			DB.close(pstmt);
 			pstmt = null;
 		}
-		
+
 		return rowSet;
 	}
 }	//	MPrintFormat
