+++ de.metas.adempiere.adempiere/base/src/test/java/org/compiere/util/UtilTest.java
@@ -6,6 +6,8 @@ import org.adempiere.exceptions.AdempiereException;
 import org.junit.Assert;
 import org.junit.Test;
 
+import de.metas.util.StringUtils;
+
 /*
  * #%L
  * de.metas.adempiere.adempiere.base
@@ -83,7 +85,7 @@ public class UtilTest
 		final String emptyString = "";
 		final String zeroString = "0000000000";
 
-		assertThat(Util.lpadZero(emptyString, 10, "This Is An Empty String")).isEqualTo(zeroString);
+		assertThat(StringUtils.lpadZero(emptyString, 10, "This Is An Empty String")).isEqualTo(zeroString);
 	}
 
 	@Test(expected = AdempiereException.class)
@@ -91,7 +93,7 @@ public class UtilTest
 	{
 		final String value = "123456789";
 
-		Util.lpadZero(value, 2, "This Is An Empty String");
+		StringUtils.lpadZero(value, 2, "This Is An Empty String");
 
 	}
 
@@ -101,7 +103,7 @@ public class UtilTest
 		final String value = "123";
 		final String zeroString = "000123";
 
-		assertThat(Util.lpadZero(value, 6, "This Is An Empty String")).isEqualTo(zeroString);
+		assertThat(StringUtils.lpadZero(value, 6, "This Is An Empty String")).isEqualTo(zeroString);
 	}
 
 	@Test
@@ -110,7 +112,7 @@ public class UtilTest
 		final String emptyString = "";
 		final String zeroString = "0000000000";
 
-		assertThat(Util.rpadZero(emptyString, 10, "This Is An Empty String")).isEqualTo(zeroString);
+		assertThat(StringUtils.rpadZero(emptyString, 10, "This Is An Empty String")).isEqualTo(zeroString);
 	}
 
 	@Test(expected = AdempiereException.class)
@@ -118,7 +120,7 @@ public class UtilTest
 	{
 		final String value = "123456789";
 
-		Util.rpadZero(value, 2, "This Is An Empty String");
+		StringUtils.rpadZero(value, 2, "This Is An Empty String");
 
 	}
 
@@ -128,7 +130,7 @@ public class UtilTest
 		final String value = "123";
 		final String zeroString = "123000";
 
-		assertThat(Util.rpadZero(value, 6, "This Is An Empty String")).isEqualTo(zeroString);
+		assertThat(StringUtils.rpadZero(value, 6, "This Is An Empty String")).isEqualTo(zeroString);
 	}
 
 	@Test
@@ -154,7 +156,7 @@ public class UtilTest
 
 		// 2 minimums, 1 greater
 		assertThat(Util.getMinimumOfThree(minimumNumber, minimumNumber, mediumNumber)).isEqualTo(minimumNumber);
-		
+
 	}
 
 }
