+++ de.metas.adempiere.adempiere/client/src/main/java/org/adempiere/ui/notifications/NotificationItemPanel.java
@@ -10,12 +10,12 @@ package org.adempiere.ui.notifications;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -36,18 +36,17 @@ import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.UIManager;
 
-import net.miginfocom.swing.MigLayout;
-
 import org.adempiere.plaf.SwingEventNotifierUI;
 import org.compiere.Adempiere;
-import org.compiere.util.Util;
 
 import de.metas.adempiere.gui.ADHyperlinkHandler;
 import de.metas.util.Check;
+import de.metas.util.StringUtils;
+import net.miginfocom.swing.MigLayout;
 
 /**
  * UI component for {@link NotificationItem}.
- * 
+ *
  * @author tsa
  *
  */
@@ -176,7 +175,7 @@ class NotificationItemPanel extends JPanel
 
 		if (!Check.isEmpty(item.getSummary(), true))
 		{
-			summaryText.setText(Util.maskHTML(item.getSummary().trim(), false));
+			summaryText.setText(StringUtils.maskHTML(item.getSummary().trim(), false));
 			visible = true;
 		}
 		else
