+++ de.metas.handlingunits.client/src/main/java/de/metas/handlingunits/client/terminal/empties/model/EmptiesKey.java
@@ -10,12 +10,12 @@ package de.metas.handlingunits.client.terminal.empties.model;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program.  If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -29,7 +29,6 @@ import java.util.List;
 
 import org.compiere.model.I_C_BPartner;
 import org.compiere.util.KeyNamePair;
-import org.compiere.util.Util;
 
 import de.metas.adempiere.form.terminal.ITerminalKey;
 import de.metas.adempiere.form.terminal.TerminalKey;
@@ -39,6 +38,7 @@ import de.metas.handlingunits.model.I_M_HU_PI;
 import de.metas.handlingunits.model.I_M_HU_PackingMaterial;
 import de.metas.util.Check;
 import de.metas.util.Services;
+import de.metas.util.StringUtils;
 
 /**
  * Empties Key
@@ -157,7 +157,7 @@ public class EmptiesKey extends TerminalKey
 	{
 		final StringBuilder name = new StringBuilder();
 
-		name.append(Util.maskHTML(piName));
+		name.append(StringUtils.maskHTML(piName));
 
 		final int qty = getQty();
 		final String qtyStr = qty > 0 ? String.valueOf(qty) : "0";
